+++ src/routes/create/post/+page.svelte
@@ -1,10 +1,19 @@
 <script lang="ts">
-    import { goto } from '$app/navigation'
-    import PostForm from '$lib/components/lemmy/post/PostForm.svelte'
-    import Card from '$lib/components/ui/Card.svelte'
-    import { profile } from '$lib/auth.js'
-    import { onMount } from 'svelte'
+    import type { Community } from 'lemmy-js-client'    
+    import { clearLastSeenCommunity, getLastSeenCommunity } from '$lib/components/lemmy/community/helpers.js';
     import { getSessionStorage, setSessionStorage } from '$lib/session.js'
+    import { goto } from '$app/navigation'
+    import { onMount } from 'svelte'    
+    import { profile } from '$lib/auth.js'
+    import { site } from '$lib/lemmy'
+    
+    
+    import PostForm from '$lib/components/lemmy/post/PostForm.svelte'
+    import SubNavbar from '$lib/components/ui/subnavbar/SubNavbar.svelte';
+    import MainContentArea from '$lib/components/ui/containers/MainContentArea.svelte';
+    import FeedContainer from '$lib/components/ui/containers/FeedContainer.svelte';
+    import SiteCard from '$lib/components/lemmy/SiteCard.svelte';
+    
 
     export let data
 
@@ -14,25 +23,30 @@
         }
     })
 
-    let community = getSessionStorage('lastSeenCommunity') as
-        | { id: number; name: string }
-        | undefined
-
-    setSessionStorage('lastSeenCommunity', undefined)
-
     let draft = getSessionStorage('postDraft') as any
 </script>
 
 <svelte:head>
-    <title>Create post</title>
+    <title>Create Post</title>
 </svelte:head>
 
-<div class="w-full max-w-5xl mx-auto">
-    <PostForm
-        data={data.crosspost ? draft : undefined}
-        passedCommunity={community}
-        on:submit={(e) => goto(`/post/${e.detail.post.id}`)}
-    >
-    <h1 class="text-2xl font-bold" slot="formtitle">Create Post</h1>
-    </PostForm>
-</div>
+<SubNavbar home back toggleMargins toggleCommunitySidebar />
+
+<MainContentArea>
+    <FeedContainer>
+        
+            <PostForm crosspostData={data.crosspost ? draft : undefined}
+                on:submit={(e) => goto(`/post/${e.detail.post.id}`)}
+            >
+            <h1 class="text-2xl font-bold" slot="formtitle">Create Post</h1>
+            </PostForm>
+        
+    </FeedContainer>
+
+    <div class="h-full" slot="right-panel">
+        {#if $site }
+            <SiteCard site={$site.site_view} taglines={$site.taglines} admins={$site.admins} version={$site.version} />
+        {/if}
+
+    </div>
+</MainContentArea>